refactor: LottoCountDto -> LottoCount
