+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.dto.LottoCountDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,6 +10,6 @@ public class LottoMachineTest {
 	@DisplayName("주어진 개수만큼 자동 로또 사기")
 	void buyAutoLotto() {
 		LottoMachine lottoMachine = new LottoMachine();
-		assertThat(lottoMachine.makeRandomLottos(new LottoCountDto(6000)).size()).isEqualTo(6);
+		assertThat(lottoMachine.makeRandomLottos(new LottoCount(6000)).size()).isEqualTo(6);
 	}
 }
